refactor: 폴더 구조 재정의
