+++ src/core/mocks/browser.js
