+++ src/core/routes/index.ts
