+++ src/core/routes/router.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from 'react-router-dom';
-import App from '../App';
+import App from '../../App';
 
 export const router = createBrowserRouter([
   {
