+++ src/shared/components/Avatar/index.tsx
