+++ src/shared/components/Footer/index.tsx
