+++ src/shared/components/Skeleton/index.tsx
