+++ src/shared/components/Slider/index.tsx
